コード微調整と天気画像の追加

+++ database/migrations/2021_04_05_184644_create_document_question_table.php
@@ -15,8 +15,8 @@ class CreateDocumentQuestionTable extends Migration
     {
         Schema::create('document_question', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('document_id');
-            $table->integer('question_id');
+            $table->unsignedBigInteger('document_id');
+            $table->unsignedBigInteger('question_id');
             $table->timestamps();
         });
     }
