+++ database/migrations/2021_05_25_095152_create_question_user_table.php
@@ -15,8 +15,8 @@ class CreateQuestionUserTable extends Migration
     {
         Schema::create('question_user', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('question_id');
-            $table->integer('user_id');
+            $table->unsignedBigInteger('question_id');
+            $table->unsignedBigInteger('user_id');
             $table->timestamps();
         });
     }
